+++ plugins/e_menu/e_menu/src/obbit.tcl
@@ -5,7 +5,7 @@
 # The ObjectProperty class allows to mix-in into
 # an object the getter and setter of properties.
 #
-# The ThemingManager class allows to change the ttk widgets' style.
+# The ObjectTheming class allows to change the ttk widgets' style.
 # For now it's only a bit of what should be, it needs to be enhanced a lot.
 #
 # The ObjectUtils class provides methods to extract option values... and
@@ -107,7 +107,7 @@ oo::class create apave::ObjectTheming {
     if {![catch {set oldval [ttk::style $oper $ts $opt]}]} {
       catch {ttk::style $oper $ts $opt $val}
       if {$oldval=="" && $oper=="configure"} {
-        switch $opt {
+        switch -- $opt {
           -foreground - -background {
             set oldval [ttk::style $oper . $opt]
           }
@@ -308,9 +308,6 @@ oo::class create apave::ObjectTheming {
         my Ttk_style map $typ $v1 [list {*}$v2]
       }
     }
-    # at last, separate widget types
-    ttk::style configure TButton \
-      -anchor center -width -11 -padding 3 -relief raised -borderwidth 2
     set _OT_Init 0
 
   }
@@ -322,7 +319,7 @@ oo::class create apave::ObjectTheming {
 
     classvar _OT_Init _OT_Opts _OT_OldOpts
     foreach {oper ts opt val} $_OT_OldOpts {
-      switch $oper {
+      switch -- $oper {
         map -
         configure {
           ttk::style $oper $ts $opt $val
@@ -384,7 +381,7 @@ oo::class create apave::ObjectTheming {
   method NonThemedWidgets {selector} {
 
     classvar _OT_Init _OT_Opts _OT_OldOpts
-    switch $selector {
+    switch -- $selector {
       entry {
         return [list entry text listbox spinbox]
       }
@@ -433,7 +430,7 @@ oo::class create apave::ObjectTheming {
       set disopt ""
       if {[info exist _OT_Opts(disabled,0)]} {
         set typ [string range [lindex [split $typ .] end] 0 2]
-        switch $typ {
+        switch -- $typ {
           frA - lfR {
             append disopt " " $_OT_Opts(disabled,2)
           }