+++ plugins/add_shortcut/release_notes.md
@@ -1,9 +1,9 @@
 # Last changes:
 
 
-Version `1.9 (20 Apr'20)`
+Version `1.9 (25 Apr'20)`
 
-  - apave v2.5
+  - apave v2.6.2
   - .md, .markdown to run with F9